+++ src/core/builtins/builtins_optionals.ml
@@ -22,6 +22,7 @@ let () =
       ("graphics", Graphics_option.enabled);
       ("gstreamer", Gstreamer_option.enabled);
       ("inotify", Inotify_option.enabled);
+      ("imagelib", Imagelib_option.enabled);
       ("irc", Irc_option.enabled);
       ("ladspa", Ladspa_option.enabled);
       ("lame", Lame_option.enabled);